requirements.txt
torch
numpy
python-dotenv
git+https://github.com/Boritech-Solutions/PyEarthworm